コンポーネント単位のリファクタリング
